fix: changed s3 code
